git-svn-id: svn+ssh://svn.msr-inria.inria.fr/var/lib/svn/repository/tla/trunk/tla2-inria@25234 76a6fc44-f60b-0410-a9a8-e67b0e8fc65c

+++ tlatools/test-long/tlc2/tool/fp/DiskFPSetTest.java
@@ -8,7 +8,7 @@ public class DiskFPSetTest extends FPSetTest {
 	 * @see tlc2.tool.fp.AbstractFPSetTest#getFPSet(int)
 	 */
 	@Override
-	protected FPSet getFPSet(int freeMemory) throws IOException {
+	protected FPSet getFPSet(long freeMemory) throws IOException {
 		return new DiskFPSet(freeMemory);
 	}
 }
